coachvirtualbackend/coachvirtualback/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
chardet==4.0.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==1.16.0
gunicorn==23.0.0
huami_token==0.7.0
idna==2.10
psycopg==3.2.12
psycopg-binary==3.2.12
PyJWT==2.10.1
pymongo==3.12.0
python-decouple==3.8
requests==2.25.1
sqlparse==0.5.3
stripe==13.2.0
types-requests==2.25.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
whitenoise==6.11.0
